styling and adding new functionalities

+++ events/views.py
@@ -3,12 +3,14 @@ import io
 from datetime import datetime
 from django.shortcuts import render, redirect
 from django.http import HttpResponseRedirect
+from django.contrib.auth.models import User
 from django.http import HttpResponse
 from django.contrib import messages
 from calendar import HTMLCalendar
 from .models import Event, Venue
-from .forms import EventForms, VenueForm
+from .forms import EventFormAdmin,EventFormUser, VenueForm
 from django.http import FileResponse
+from django.contrib.auth import logout
 
 #it is necesserly  to install reportlab
 from reportlab.pdfgen import canvas
@@ -55,7 +57,7 @@ def all_events(request):
     page = request.GET.get("page")
     events = paginator.get_page(page) 
     
-    return render (request,"event_list.html",{
+    return render (request,"event/event_list.html",{
         "events_list":events_list,
         "events":events
     })
@@ -63,7 +65,7 @@ def all_events(request):
 def event_detail(request,id):
     if request.method == "GET":
         event = Event.objects.get(pk = id)
-        return render (request,"event_detail.html",{"event":event})
+        return render (request,"events/event_detail.html",{"event":event})
     
 def add_venue(request):
     submitted = False
@@ -81,29 +83,46 @@ def add_venue(request):
         form = VenueForm()
         if "submitted" in request.GET:
             submitted = True
-        return render (request,"create_venue.html",{"form":form,
+        return render (request,"venue/create_venue.html",{"form":form,
                                                     "submitted":submitted} )
 
 def all_venues(request):
     venues = Venue.objects.all().order_by("name") #   ?<- order by random 
-    return render(request,"venue.html",{"venues":venues})
+    return render(request,"venue/venue.html",{"venues":venues})
 
 #venue detail
 def show_venue(request,id):
-    venue = Venue.objects.all().filter(pk = id)
-    return render(request,"venue_detail.html",{"venues":venue})
+    venue = Venue.objects.get(pk = id)
+    venue_owner = User.objects.get(pk = venue.owner)
+    return render(request,"/venue/venue_detail.html",{"venue":venue,
+                                               "venue_owner":venue_owner})
     
 def search_venues(request):
     if request.method == "POST":
         searched = request.POST["searched"]
         venues = Venue.objects.filter(name__contains = searched)
         
-        return render (request,"search_venue.html",{"searched":searched,
+        return render (request,"venue/search_venue.html",{"searched":searched,
                                                     "venues":venues}) 
     else:
-        return render (request,"search_venue.html",{})
+        return render (request,"venue/search_venue.html",{})
     
-    
+def search_events(request):
+    if request.method == "POST":
+        searched = request.POST.get("searched")
+        events = Event.objects.filter(name__contains = searched)
+            
+        if events:
+            return render(request,"event/search_events.html",{"searched":searched,
+                                                      "events":events})
+            
+        else:
+            messages.info(request,"Didn`t find any of searched! Try Again!")
+            return render(request,"event/search_events.html")
+    else:
+       return render(request,"event/search_events.html",{}) 
+   
+   
 def update_venue (request,id):
     updated = False 
     if request.method == "POST":
@@ -117,7 +136,7 @@ def update_venue (request,id):
       venue_form = VenueForm(request.POST or None, instance = venue)
       if "updated" in request.GET:
           updated = True
-      return render (request,"update_venue.html",{"venue":venue,
+      return render (request,"venue/update_venue.html",{"venue":venue,
                                                   "form":venue_form,
                                                   "updated":updated})
       
@@ -125,19 +144,35 @@ def update_venue (request,id):
 def add_event(request):
     submitted = False
     if request.method == "POST":
-        form = EventForms(request.POST)
-
-        if form.is_valid():
-           form.save()
-           return HttpResponseRedirect("/add-event?submitted=True")
+        if request.user.is_superuser:    
+            form = EventFormAdmin(request.POST)
+            if form.is_valid():
+               form.save()
+               return HttpResponseRedirect("/add-event?submitted=True")
+            else:
+                #messages.info(request,"It is possible that you have entered wrong date format!")
+                return render(request,"event/create_events.html",{"form":form})
+      
         else:
-            #messages.info(request,"It is possible that you have entered wrong date format!")
-            return render(request,"create_events.html",{"form":form})
+            form = EventFormUser(request.POST)
+            if form.is_valid():
+               form = form.save(commit=False)
+               form.manager = request.user
+               form.save()
+               return HttpResponseRedirect("/add-event?submitted=True")
+            else:
+                #messages.info(request,"It is possible that you have entered wrong date format!")
+                return render(request,"event/create_events.html",{"form":form})
+      
     else:
-        form = EventForms()
+        #just going to the page, not submmit
+        if request.user.is_superuser:       
+             form = EventFormAdmin()
+        else:
+            form = EventFormUser()
         if "submitted" in request.GET:
             submitted = True
-        return render(request,"create_events.html",{"form":form,
+        return render(request,"event/create_events.html",{"form":form,
                                                     "submitted":submitted})  
     
 
@@ -145,25 +180,38 @@ def update_event(request,id):
     updated = False
     if request.method == "POST":
         event = Event.objects.get(pk = id)
-        form  = EventForms(request.POST or None,instance = event)
+        if request.user.is_superuser:
+            form  = EventFormAdmin(request.POST or None,instance = event)
+        else:
+            form  = EventFormUser(request.POST or None,instance = event)
+        #check form validation and save it     
         if form.is_valid():
            form.save()
            return HttpResponseRedirect(f"/update-event/{id}?updated=True")
         else:
-            return render(request,"update_events.html",{"form":form})
+            messages.info(request,"Something went wrong! Try again!")
+            return render(request,"event/update_events.html",{"form":form})
     else:
        event = Event.objects.get(pk = id)
-       form  = EventForms(request.POST or None,instance = event) 
+       if request.user.is_superuser: 
+          form  = EventFormAdmin(request.POST or None,instance = event) 
+       else:
+         form  = EventFormUser(request.POST or None,instance = event) 
        if "updated" in request.GET:
            updated = True
-       return render(request,"update_event.html",{"form":form,
+       return render(request,"event/update_event.html",{"form":form,
                                                   "updated":updated,
                                                   "event":event})
 
 def delete_event(request,id):
     event = Event.objects.get(pk = id)
-    event.delete()
-    return redirect("all_events")
+    if request.user == event.manager:
+       event.delete()
+       messages.success(request,"Sucessfully deleted!")
+       return redirect("all_events")
+    else:
+        messages.info(request,"You are not able to delete this event!")
+        return redirect("all_events")
     
 def delete_venue(request,id):
     venue = Venue.objects.get(pk = id)
@@ -254,4 +302,22 @@ def event_pdf (request,id):
     byte.seek(0)
     
     return FileResponse(byte,as_attachment=True,filename = f"{event.name}.pdf")
-     
+
+
+def my_events(request,manager):  
+    if request.user.is_authenticated:    
+      manager = request.user
+      my_event = Event.objects.filter(manager = manager)
+      if my_event:
+        paginator = Paginator(Event.objects.filter(manager = manager).order_by("event_date"),1)
+        page = request.GET.get("page")
+        events = paginator.get_page(page)
+        return render(request,"event/my_events.html",{"events":events})
+      else: 
+          messages.info(request,"You don`t have any event! If you want to add one, click on button below!\n")
+          return render(request,"event/my_events.html",{"my_event":my_event}) 
+    else:
+        logout(request)
+        messages.info(request,"Please log in to see page!")
+        return render(request,"event/my_events.html")
+
